issue number 155 fixed

+++ SocialCAM/Controlers/ContactWizardAndAbout/ContactImportVC.swift
@@ -69,6 +69,8 @@ class ContactImportVC: UIViewController, UITableViewDelegate, UITableViewDataSou
     @IBOutlet weak var socialShareView: UIView!
     @IBOutlet weak var businessDashboardView: UIView!
     
+    @IBOutlet weak var deleteContactConfirmationView: UIView!
+    @IBOutlet weak var deleteContactDoNotShowButton: UIButton!
     @IBOutlet weak var filterOptionView: UIView!
     var loadingStatus = false
     let blueColor1 = UIColor(red: 0/255, green: 125/255, blue: 255/255, alpha: 1.0)
@@ -277,6 +279,8 @@ class ContactImportVC: UIViewController, UITableViewDelegate, UITableViewDataSou
         self.manualEmailView.dropShadow()
         self.socialShareView.dropShadow()
         self.businessDashboardView.dropShadow()
+        
+        previewImageview.contentMode = .scaleAspectFit
        
     }
     override func viewWillAppear(_ animated: Bool) {
@@ -1090,7 +1094,7 @@ class ContactImportVC: UIViewController, UITableViewDelegate, UITableViewDataSou
             DispatchQueue.main.async {
                 self.lblpreviewUrl.text = link
                 self.previewImageview.layer.cornerRadius = self.previewImageview.bounds.width / 2
-                self.previewImageview.contentMode = .scaleAspectFill
+            
 //                self.lblpreviewText.text = self.txtLinkWithCheckOut
             }
 //            if ogData.imageUrl != nil {
@@ -1501,17 +1505,23 @@ class ContactImportVC: UIViewController, UITableViewDelegate, UITableViewDataSou
             return UIContextMenuConfiguration(identifier: indexPath as NSIndexPath, previewProvider: nil) { _ in
                 return UIMenu(title: "", children: [
                     UIAction(title: "Delete", image: UIImage(systemName: "")) { action in
-                        if tableView == self.emailContactTableView{
-                            let contact = self.emailContacts[indexPath.row]
-                            print(contact.email ?? "")
-                            self.deleteContact(contact: contact, isEmail: true, index: indexPath.row)
-                            
-                        }else if tableView == self.contactTableView{
-                            let contact = self.mobileContacts[indexPath.row]
-                            print(contact.mobile ?? "")
-                            self.deleteContact(contact: contact, isEmail: false, index: indexPath.row)
-                        }
                         
+                        if Defaults.shared.isDoNotShowAgainDeleteContactPopup{
+                            self.deleteContactConfirmationView.isHidden = true
+                            if tableView == self.emailContactTableView{
+                                let contact = self.emailContacts[indexPath.row]
+                                print(contact.email ?? "")
+                                self.deleteContact(contact: contact, isEmail: true, index: indexPath.row)
+                                
+                            }else if tableView == self.contactTableView{
+                                let contact = self.mobileContacts[indexPath.row]
+                                print(contact.mobile ?? "")
+                                self.deleteContact(contact: contact, isEmail: false, index: indexPath.row)
+                            }
+                        }else{
+                            self.deleteContactConfirmationView.tag = indexPath.row
+                            self.deleteContactConfirmationView.isHidden = false
+                        }
                     },UIAction(title: "Edit", image: UIImage(systemName: "")) { action in
                         if tableView == self.emailContactTableView{
                             let contact = self.emailContacts[indexPath.row]
@@ -1985,15 +1995,27 @@ class ContactImportVC: UIViewController, UITableViewDelegate, UITableViewDataSou
         
         controller.dismiss(animated: true, completion: nil)
     }
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
+   
+    @IBAction func deleteContactDoNotShowClick(_ sender: UIButton) {
+        
+        deleteContactDoNotShowButton.isSelected = !deleteContactDoNotShowButton.isSelected
+        Defaults.shared.isDoNotShowAgainDeleteContactPopup = deleteContactDoNotShowButton.isSelected
     }
-    */
+    @IBAction func deleteContactYesClick(_ sender: UIButton) {
+        deleteContactConfirmationView.isHidden = true
+        if selectedContactType == ContactType.mobile{
+            let contact = self.mobileContacts[deleteContactConfirmationView.tag]
+            print(contact.mobile ?? "")
+            self.deleteContact(contact: contact, isEmail: false, index: deleteContactConfirmationView.tag)
+        }else{
+            let contact = self.emailContacts[deleteContactConfirmationView.tag]
+            print(contact.mobile ?? "")
+            self.deleteContact(contact: contact, isEmail: true, index: deleteContactConfirmationView.tag)
+        }
+    }
+    @IBAction func deleteContactNoClick(_ sender: UIButton) {
+        deleteContactConfirmationView.isHidden = true
+    }
 
 }
 extension ContactImportVC:UIScrollViewDelegate{
